requirements.txt
# PathoGen - Histopathology Image Inpainting
# Core Dependencies

# PyTorch and related
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0

# Diffusion models
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.24.0

# Configuration
hydra-core>=1.3.0
omegaconf>=2.3.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0

# CLI
click>=8.1.0

# Progress bars
tqdm>=4.65.0

# Additional utilities
scipy>=1.11.0
huggingface-hub>=0.19.0

